Replace / edit badge icon via admin tool

+++ modules/admin/certbadge.php
@@ -29,12 +29,12 @@ $navigation[] = array('url' => 'index.php', 'name' => $langAdmin);
 
 $tool_content .= action_bar(array(
         array('title' => "$langAddNewCertTemplate",
-              'url' => "$_SERVER[SCRIPT_NAME]?add=cert",
+              'url' => "$_SERVER[SCRIPT_NAME]?action=add_cert",
               'icon' => 'fa-plus-circle',
               'level' => 'primary-label',
               'button-class' => 'btn-success'),
         array('title' => "$langAddNewBadgeTemplate",
-              'url' => "$_SERVER[SCRIPT_NAME]?add=badge",
+              'url' => "$_SERVER[SCRIPT_NAME]?action=add_badge",
               'icon' => 'fa-plus-circle',
               'level' => 'primary-label',
               'button-class' => 'btn-success'),
@@ -59,7 +59,7 @@ if (isset($_GET['del_cert'])) { // delete certificate template
     }    
 }
 
-if (isset($_POST['submit_cert_template'])) {
+if (isset($_POST['submit_cert_template'])) { // insert certificate template
     $filename = $_FILES['filename']['name'];
     if (move_uploaded_file($_FILES['filename']['tmp_name'], "$webDir" . CERT_TEMPLATE_PATH . "$filename")) {
         Database::get()->querySingle("INSERT INTO certificate_template SET 
@@ -68,20 +68,40 @@ if (isset($_POST['submit_cert_template'])) {
                                         filename = ?s", $_POST['name'], $_POST['description'], $filename);
         Session::Messages($langDownloadEnd, 'alert-success');
     }
-} elseif (isset($_POST['submit_badge_icon'])) {
-    $filename = $_FILES['icon']['name'];
-    if (move_uploaded_file($_FILES['icon']['tmp_name'], "$webDir" . BADGE_TEMPLATE_PATH . "$filename")) {
-        Database::get()->querySingle("INSERT INTO badge_icon SET 
-                                        name = ?s, 
+} elseif (isset($_POST['submit_badge_icon'])) { // insert / update badge icon
+    if (isset($_POST['badge_id'])) {        
+        if ($_FILES['icon']['size'] > 0) { // replace file if needed
+            $filename = $_FILES['icon']['name'];
+            if (move_uploaded_file($_FILES['icon']['tmp_name'], "$webDir" . BADGE_TEMPLATE_PATH . "$filename")) {
+                $old_icon = Database::get()->querySingle("SELECT filename FROM badge_icon WHERE id = ?d", $_POST['badge_id'])->filename;
+                unlink($webDir . BADGE_TEMPLATE_PATH . $old_icon); // delete old icon
+                Database::get()->querySingle("UPDATE badge_icon SET
+                                        name = ?s,
                                         description = ?s,
-                                        filename = ?s", $_POST['name'], $_POST['description'], $filename);
-        Session::Messages($langDownloadEnd, 'alert-success');
+                                        filename = ?s
+                                       WHERE id = ?d", $_POST['name'], $_POST['description'], $filename, $_POST['badge_id']);
+            }
+        } else {
+            Database::get()->querySingle("UPDATE badge_icon SET
+                                        name = ?s,
+                                        description = ?s
+                                       WHERE id = ?d", $_POST['name'], $_POST['description'], $_POST['badge_id']);
+        }
+    } else {
+        $filename = $_FILES['icon']['name'];
+        if (move_uploaded_file($_FILES['icon']['tmp_name'], "$webDir" . BADGE_TEMPLATE_PATH . "$filename")) {
+            Database::get()->querySingle("INSERT INTO badge_icon SET 
+                                            name = ?s, 
+                                            description = ?s,
+                                            filename = ?s", $_POST['name'], $_POST['description'], $filename);
+            Session::Messages($langDownloadEnd, 'alert-success');
+        }
     }
 }
 
 // display forms
-if (isset($_GET['add'])) {
-    if ($_GET['add'] == 'cert') { // add certificate template
+if (isset($_GET['action'])) {
+    if (($_GET['action'] == 'add_cert') or ($_GET['action'] == 'edit_cert')) { // add certificate template
         $tool_content .= "<div class='row'>
                         <div class='col-md-12'>
                         <div class='form-wrapper'>
@@ -114,29 +134,50 @@ if (isset($_GET['add'])) {
                             </div>
                         </div>
                     </div>";
-    } elseif ($_GET['add'] == 'badge') { // add badge icons
+    } elseif (($_GET['action'] == 'add_badge') or  ($_GET['action'] == 'edit_badge')) { // add badge icons
+        $badge_name = $badge_description = $badge_hidden_id = '';
+        if (isset($_GET['bid'])) {
+            $badge_id = $_GET['bid'];
+            $badge_data = Database::get()->querySingle("SELECT * FROM badge_icon WHERE id = ?d", $badge_id);
+            $badge_name = $badge_data->name;
+            $badge_description = $badge_data->description;
+            $badge_hidden_id = "<input type='hidden' name='badge_id' value='$badge_id'>";
+        }
+ 
         $tool_content .= "<div class='row'>
                         <div class='col-md-12'>
                         <div class='form-wrapper'>
-                        <form class='form-horizontal' role='form' action='$_SERVER[SCRIPT_NAME]' method='post' enctype='multipart/form-data'>
-                            <div class='form-group'>
+                        <form class='form-horizontal' role='form' action='$_SERVER[SCRIPT_NAME]' method='post' enctype='multipart/form-data'>";
+                        if (isset($_GET['bid'])) {
+                            $icon_link = $urlServer . BADGE_TEMPLATE_PATH . "$badge_data->filename";
+                            $tool_content .= "<div class='form-group'>
+                                <label class='col-sm-2 control-label'>$langReplace:</label>
+                                <div class='col-sm-10'>
+                                    <img src='$icon_link' width='60' height='60'>
+                                    <input type='file' name='icon' value=''>
+                                </div>
+                            </div>";
+                        } else {
+                            $tool_content .= "<div class='form-group'>
                                 <label class='col-sm-2 control-label'>$langIcon:</label>
                                 <div class='col-sm-10'>
                                     <input type='file' name='icon' value=''>
                                 </div>
-                            </div>
-                            <div class='form-group'>
+                            </div>";
+                        }
+                        $tool_content .= "<div class='form-group'>
                                 <label class='col-sm-2 control-label'>$langName:</label>
                                 <div class='col-sm-10'>
-                                    <input type='text' class='form-control' name='name' value=''>
+                                    <input type='text' class='form-control' name='name' value='$badge_name'>
                                 </div>
                             </div>
                             <div class='form-group'>
                             <label for='description' class='col-sm-2 control-label'>$langDescription: </label>
                                 <div class='col-sm-10'>
-                                    " . rich_text_editor('description', 2, 60, '') . "
+                                    " . rich_text_editor('description', 2, 60, $badge_description) . "
                                 </div>
                             </div>
+                            $badge_hidden_id
                             <div class='form-group'>
                                 <div class='col-xs-offset-2 col-xs-10'>
                                     <button class='btn btn-primary' type ='submit' name='submit_badge_icon'>$langUpload</button>
@@ -156,7 +197,7 @@ if (isset($_GET['add'])) {
                         <tr class='list-header'>
                             <th>$langTitle</th>
                             <th>$langDescription</th>                            
-                            <th width='60' class='text-center'><i class='fa fa-cogs'></i></th>
+                            <th class='text-center'><i class='fa fa-cogs'></i></th>
                         </tr>";
                 foreach ($sql1 as $cert_data) {
                     $template_link = $urlServer . CERT_TEMPLATE_PATH ."$cert_data->filename";
@@ -168,11 +209,15 @@ if (isset($_GET['add'])) {
                                     'icon' => 'fa-times',
                                     'url' => "$_SERVER[SCRIPT_NAME]?del_cert=$cert_data->id",
                                     'confirm' => $langConfirmDelete,
-                                    'class' => 'delete'))).
+                                    'class' => 'delete'),
+                                array('title' => $langEdit,
+                                    'icon' => 'fa-edit',
+                                    'url' => "$_SERVER[SCRIPT_NAME]?action=edit_cert&amp;cid=$cert_data->id"
+                                    ))).
                             "</td></tr>";
                 }
-                $tool_content .= "</table>";
-                $tool_content .= "</div></div></div>";
+    $tool_content .= "</table>";
+    $tool_content .= "</div></div></div>";
     
     $sql2 = Database::get()->queryArray("SELECT * FROM badge_icon");
     
@@ -183,7 +228,7 @@ if (isset($_GET['add'])) {
                             <th>$langTitle</th>
                             <th>$langDescription</th>
                             <th width='70' class='text-center'>$langIcon</th>
-                            <th width='60' class='text-center'><i class='fa fa-cogs'></i></th>
+                            <th class='text-center'><i class='fa fa-cogs'></i></th>
                         </tr>";
                 foreach ($sql2 as $badge_data) {
                     $icon_link = $urlServer . BADGE_TEMPLATE_PATH ."$badge_data->filename";
@@ -191,7 +236,11 @@ if (isset($_GET['add'])) {
                                       <td>" . ellipsize_html($badge_data->description, 100) . "</td>
                                       <td class='text-center'><img src='$icon_link' width='50' height='50'></td>";
                     $tool_content .= "<td class='text-center option-btn-cell'>".
-                            action_button(array(
+                            action_button(array(                                
+                                array('title' => $langEdit,
+                                    'icon' => 'fa-edit',
+                                    'url' => "$_SERVER[SCRIPT_NAME]?action=edit_badge&amp;bid=$badge_data->id"
+                                    ),
                                 array('title' => $langDelete,
                                     'icon' => 'fa-times',
                                     'url' => "$_SERVER[SCRIPT_NAME]?del_badge=$badge_data->id",
@@ -199,8 +248,8 @@ if (isset($_GET['add'])) {
                                     'class' => 'delete'))).
                             "</td></tr>";
                 }
-                $tool_content .= "</table>";
-                $tool_content .= "</div></div></div>";
+    $tool_content .= "</table>";
+    $tool_content .= "</div></div></div>";
 }
 
 draw($tool_content, 3);
